Move Java code to action based system
