+++ src/android/java/com/pspdfkit/cordova/CordovaPdfActivity.java
@@ -1,6 +1,5 @@
 package com.pspdfkit.cordova;
 
-import com.pspdfkit.ui.PdfActivity;
-
-public class CordovaPdfActivity extends PdfActivity {
+public interface CordovaPdfActivity {
+    void onDismissRequested();
 }
