+++ src/android/java/com/pspdfkit/cordova/action/Action.java
@@ -1,0 +1,12 @@
+package com.pspdfkit.cordova.action;
+
+import org.apache.cordova.CallbackContext;
+import org.json.JSONArray;
+
+import androidx.annotation.NonNull;
+
+public interface Action {
+  @NonNull String getName();
+
+  boolean exec(JSONArray args, CallbackContext callbackContext);
+}
